Refactor pi-codex provider structure (#32)
* Remove applied pi-codex changeset

* Add pi-codex provider characterization tests

* Expand pi-codex provider stream tests

* Split pi-codex provider helpers

* Extract pi-codex websocket transport

* Split pi-codex websocket internals

* Restore pi-codex provider type exports
