+++ packages/pi-codex-conversion/tests/openai-codex-custom-provider.test.ts
@@ -4,6 +4,8 @@ import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
 import {
+	IMAGE_SAVE_DISPLAY_MESSAGE_TYPE,
+	WEB_SEARCH_ACTIVITY_MESSAGE_TYPE,
 	buildProviderErrorMessage,
 	buildRequestBody,
 	buildCachedWebSocketRequestBody,
@@ -16,8 +18,19 @@ import {
 	getOpenAICodexLatestImagePath,
 	getOpenAICodexImagePath,
 	parseSSE,
+	registerOpenAICodexCustomProvider,
 	saveOpenAICodexGeneratedImage,
 } from "../src/providers/openai-codex-custom-provider.ts";
+
+const webSearchTool = {
+	name: "web.run",
+	description: "Search the web",
+	parameters: {
+		type: "object",
+		properties: { query: { type: "string" } },
+		required: ["query"],
+	},
+} as never;
 
 const codexModel = {
 	provider: "openai-codex",
@@ -35,6 +48,47 @@ async function waitForTimers(): Promise<void> {
 	await new Promise<void>((resolve) => setTimeout(resolve, 0));
 }
 
+function fakeJwt(payload: Record<string, unknown>): string {
+	return ["header", Buffer.from(JSON.stringify(payload)).toString("base64url"), "signature"].join(".");
+}
+
+function sseResponse(events: unknown[]): Response {
+	return new Response(events.map((event) => `data: ${JSON.stringify(event)}\n\n`).join(""), {
+		status: 200,
+		headers: { "content-type": "text/event-stream" },
+	});
+}
+
+async function collectStream(stream: AsyncIterable<unknown>): Promise<unknown[]> {
+	const events: unknown[] = [];
+	for await (const event of stream) events.push(event);
+	return events;
+}
+
+function createRegisteredCodexProvider(options?: { cwd?: string | undefined }) {
+	const providers = new Map<string, { streamSimple: (...args: never[]) => AsyncIterable<unknown> }>();
+	const handlers = new Map<string, Array<(...args: never[]) => unknown>>();
+	const renderers = new Map<string, unknown>();
+	const sentMessages: Array<{ message: unknown; options: unknown }> = [];
+	const pi = {
+		registerProvider(id: string, provider: { streamSimple: (...args: never[]) => AsyncIterable<unknown> }) {
+			providers.set(id, provider);
+		},
+		on(event: string, handler: (...args: never[]) => unknown) {
+			handlers.set(event, [...(handlers.get(event) ?? []), handler]);
+		},
+		registerMessageRenderer(type: string, renderer: unknown) {
+			renderers.set(type, renderer);
+		},
+		sendMessage(message: unknown, messageOptions: unknown) {
+			sentMessages.push({ message, options: messageOptions });
+		},
+	};
+
+	registerOpenAICodexCustomProvider(pi as never, { getCurrentCwd: () => options?.cwd ?? process.cwd() });
+	return { provider: providers.get("openai-codex")!, handlers, renderers, sentMessages };
+}
+
 test("buildRequestBody sends a non-empty fallback system prompt", () => {
 	const body = buildRequestBody(codexModel, { systemPrompt: "", messages: [] });
 	assert.equal(body.instructions, "You are a helpful assistant.");
@@ -43,6 +97,183 @@ test("buildRequestBody sends a non-empty fallback system prompt", () => {
 test("buildRequestBody preserves provided system prompts", () => {
 	const body = buildRequestBody(codexModel, { systemPrompt: "Custom instructions", messages: [] });
 	assert.equal(body.instructions, "Custom instructions");
+});
+
+test("buildRequestBody keeps Codex request shape stable for common options", () => {
+	const body = buildRequestBody(
+		codexModel,
+		{
+			systemPrompt: "Instructions",
+			messages: [{ role: "user", content: "Hello" } as never],
+			tools: [webSearchTool],
+		},
+		{
+			sessionId: "session-" + "x".repeat(80),
+			serviceTier: "priority",
+			textVerbosity: "medium",
+			temperature: 0.2,
+			reasoning: "high",
+			reasoningSummary: "detailed",
+			maxTokens: 1234,
+		} as never,
+	);
+
+	assert.equal(body.model, "gpt-5.4");
+	assert.equal(body.store, false);
+	assert.equal(body.stream, true);
+	assert.equal(body.instructions, "Instructions");
+	assert.deepEqual(body.text, { verbosity: "medium" });
+	assert.equal(body.prompt_cache_key, "session-" + "x".repeat(56));
+	assert.equal(body.tool_choice, "auto");
+	assert.equal(body.parallel_tool_calls, true);
+	assert.equal(body.service_tier, "priority");
+	assert.equal(body.temperature, 0.2);
+	assert.deepEqual(body.reasoning, { effort: "high", summary: "detailed" });
+	assert.deepEqual(body.include, ["reasoning.encrypted_content", "web_search_call.action.sources", "web_search_call.results"]);
+	assert.deepEqual(body.tools, [
+		{
+			type: "function",
+			name: "web.run",
+			description: "Search the web",
+			parameters: {
+				type: "object",
+				properties: { query: { type: "string" } },
+				required: ["query"],
+			},
+			strict: null,
+		},
+	]);
+	assert.equal("max_output_tokens" in body, false, "Codex ChatGPT backend rejects max_output_tokens");
+	assert.equal("max_completion_tokens" in body, false, "Codex ChatGPT backend rejects max token aliases here");
+});
+
+test("buildRequestBody omits reasoning when Pi thinking is off", () => {
+	const body = buildRequestBody(
+		codexModel,
+		{ systemPrompt: "Instructions", messages: [] },
+		{ reasoning: "off" } as never,
+	);
+
+	assert.equal(body.reasoning, undefined);
+});
+
+test("registered Codex provider exposes provider, lifecycle handlers, and activity renderers", () => {
+	const registered = createRegisteredCodexProvider();
+
+	assert.equal(typeof registered.provider.streamSimple, "function");
+	assert.equal(registered.renderers.has(IMAGE_SAVE_DISPLAY_MESSAGE_TYPE), true);
+	assert.equal(registered.renderers.has(WEB_SEARCH_ACTIVITY_MESSAGE_TYPE), true);
+	assert.equal((registered.handlers.get("session_start") ?? []).length, 1);
+	assert.equal((registered.handlers.get("session_shutdown") ?? []).length, 1);
+	assert.equal((registered.handlers.get("agent_end") ?? []).length, 1);
+});
+
+test("registered Codex provider retries retryable SSE failures and streams the final response", async () => {
+	const originalFetch = globalThis.fetch;
+	const registered = createRegisteredCodexProvider();
+	const fetchCalls: Array<{ url: string; init: RequestInit }> = [];
+	const responseEvents = [
+		{ type: "response.created", response: { id: "resp_1" } },
+		{ type: "response.output_item.added", output_index: 0, item: { type: "message", id: "msg_1", role: "assistant", content: [] } },
+		{ type: "response.content_part.added", output_index: 0, content_index: 0, part: { type: "output_text", text: "" } },
+		{ type: "response.output_text.delta", output_index: 0, content_index: 0, delta: "Hello" },
+		{ type: "response.output_item.done", output_index: 0, item: { type: "message", id: "msg_1", role: "assistant", content: [{ type: "output_text", text: "Hello" }], status: "completed" } },
+		{ type: "response.completed", response: { id: "resp_1", status: "completed", usage: { input_tokens: 12, output_tokens: 3, total_tokens: 15, input_tokens_details: { cached_tokens: 5 } } } },
+	];
+
+	try {
+		globalThis.fetch = (async (url, init) => {
+			fetchCalls.push({ url: String(url), init: init as RequestInit });
+			return fetchCalls.length === 1
+				? new Response("temporary overloaded", { status: 500, statusText: "Server Error" })
+				: sseResponse(responseEvents);
+		}) as typeof fetch;
+
+		const onResponses: unknown[] = [];
+		const stream = registered.provider.streamSimple(
+			{ ...(codexModel as object), baseUrl: "https://chatgpt.example/backend-api", cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 } } as never,
+			{ systemPrompt: "Instructions", messages: [] } as never,
+			{
+				apiKey: fakeJwt({ "https://api.openai.com/auth": { chatgpt_account_id: "acct_1" } }),
+				transport: "sse",
+				sessionId: "session-1",
+				onResponse: (response: unknown) => onResponses.push(response),
+			} as never,
+		);
+
+		const events = await collectStream(stream);
+		const done = events.at(-1) as { type: string; message: { responseId?: string; content: Array<{ type: string; text?: string }>; usage: { input: number; cacheRead: number; output: number; totalTokens: number } } };
+
+		assert.equal(fetchCalls.length, 2);
+		assert.equal(fetchCalls[0]!.url, "https://chatgpt.example/backend-api/codex/responses");
+		assert.equal((fetchCalls[1]!.init.headers as Headers).get("session-id"), "session-1");
+		assert.equal((fetchCalls[1]!.init.headers as Headers).get("chatgpt-account-id"), "acct_1");
+		assert.equal(JSON.parse(fetchCalls[1]!.init.body as string).instructions, "Instructions");
+		assert.deepEqual(onResponses.map((response) => (response as { status: number }).status), [500, 200]);
+		assert.equal(events.some((event) => (event as { type?: string }).type === "start"), true);
+		assert.equal(events.some((event) => (event as { type?: string; delta?: string }).type === "text_delta" && (event as { delta?: string }).delta === "Hello"), true);
+		assert.equal(done.type, "done");
+		assert.equal(done.message.responseId, "resp_1");
+		assert.deepEqual(done.message.content, [{ type: "text", text: "Hello", textSignature: JSON.stringify({ v: 1, id: "msg_1" }) }]);
+		assert.deepEqual(done.message.usage, { input: 7, output: 3, cacheRead: 5, cacheWrite: 0, totalTokens: 15, cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 } });
+	} finally {
+		globalThis.fetch = originalFetch;
+	}
+});
+
+test("registered Codex provider converts non-retryable SSE errors into error events", async () => {
+	const originalFetch = globalThis.fetch;
+	const registered = createRegisteredCodexProvider();
+
+	try {
+		globalThis.fetch = (async () => new Response(JSON.stringify({ error: { message: "Bad request shape" } }), { status: 400, statusText: "Bad Request" })) as typeof fetch;
+		const events = await collectStream(registered.provider.streamSimple(
+			codexModel,
+			{ systemPrompt: "Instructions", messages: [] } as never,
+			{ apiKey: fakeJwt({ "https://api.openai.com/auth": { chatgpt_account_id: "acct_1" } }), transport: "sse" } as never,
+		));
+
+		assert.equal(events.length, 1);
+		assert.equal((events[0] as { type?: string }).type, "error");
+		assert.equal((events[0] as { error?: { errorMessage?: string } }).error?.errorMessage, "Bad request shape");
+	} finally {
+		globalThis.fetch = originalFetch;
+	}
+});
+
+test("registered Codex provider captures generated image and web search activities from SSE streams", async () => {
+	const originalFetch = globalThis.fetch;
+	const cwd = await fs.mkdtemp(path.join(os.tmpdir(), "pi-codex-provider-activity-"));
+	const registered = createRegisteredCodexProvider({ cwd });
+	const encoded = Buffer.from("png-bytes").toString("base64");
+
+	try {
+		globalThis.fetch = (async () => sseResponse([
+			{ type: "response.created", response: { id: "resp_activity" } },
+			{ type: "response.output_item.done", output_index: 0, item: { type: "image_generation_call", id: "ig_activity", result: encoded, output_format: "png", revised_prompt: "Tiny icon", status: "completed" } },
+			{ type: "response.output_item.done", output_index: 1, item: { type: "web_search_call", id: "ws_activity", status: "completed", action: { query: "docs", sources: [{ url: "https://example.com/source" }] }, results: [{ title: "Docs", url: "https://example.com/source" }] } },
+			{ type: "response.completed", response: { id: "resp_activity", status: "completed", usage: { input_tokens: 1, output_tokens: 1, total_tokens: 2 } } },
+		])) as typeof fetch;
+
+		const events = await collectStream(registered.provider.streamSimple(
+			codexModel,
+			{ systemPrompt: "Instructions", messages: [{ role: "user", content: "Draw an icon" } as never] } as never,
+			{ apiKey: fakeJwt({ "https://api.openai.com/auth": { chatgpt_account_id: "acct_1" } }), transport: "sse" } as never,
+		));
+		assert.equal((events.at(-1) as { type?: string }).type, "done");
+
+		for (const handler of registered.handlers.get("agent_end") ?? []) await handler();
+		await waitForTimers();
+
+		assert.equal(registered.sentMessages.length, 2);
+		assert.equal((registered.sentMessages[0]!.message as { customType?: string }).customType, IMAGE_SAVE_DISPLAY_MESSAGE_TYPE);
+		assert.equal((registered.sentMessages[1]!.message as { customType?: string }).customType, WEB_SEARCH_ACTIVITY_MESSAGE_TYPE);
+		assert.deepEqual(await fs.readFile(path.join(cwd, ".pi", "openai-codex-images", "ig_activity-resp_activity.png")), Buffer.from("png-bytes"));
+		assert.match((registered.sentMessages[1]!.message as { content?: string }).content ?? "", /Docs — https:\/\/example\.com\/source/);
+	} finally {
+		globalThis.fetch = originalFetch;
+		await fs.rm(cwd, { recursive: true, force: true });
+	}
 });
 
 test("buildProviderErrorMessage marks websocket failures as Pi retryable connection errors", () => {
@@ -151,6 +382,26 @@ test("cached websocket request body keeps follow-up input after drifted tool out
 	);
 });
 
+test("cached websocket request body reports explicit non-delta decisions", () => {
+	const previousBody = buildRequestBody(codexModel, { systemPrompt: "Instructions", messages: [] }, { sessionId: "session-1" });
+	previousBody.input = [{ type: "message", role: "user", content: [{ type: "input_text", text: "first" }] }];
+	const continuation = { lastRequestBody: previousBody, lastResponseId: "resp_1", lastResponseItems: [] };
+
+	assert.deepEqual(buildCachedWebSocketRequestBody(undefined, previousBody), { body: previousBody, decision: "no_continuation" });
+
+	const changedModel = { ...previousBody, model: "other-model" };
+	assert.deepEqual(buildCachedWebSocketRequestBody(continuation, changedModel), { body: changedModel, decision: "body_mismatch" });
+
+	const shorter = { ...previousBody, input: [] };
+	assert.deepEqual(buildCachedWebSocketRequestBody(continuation, shorter), { body: shorter, decision: "input_shorter_than_baseline" });
+
+	const missingPreviousResponse = { ...previousBody, input: [...previousBody.input, { type: "message", role: "user", content: [] }] };
+	assert.deepEqual(
+		buildCachedWebSocketRequestBody({ ...continuation, lastResponseId: "" }, missingPreviousResponse),
+		{ body: missingPreviousResponse, decision: "missing_previous_response_id" },
+	);
+});
+
 test("getEffectiveCodexTransport enables cached websockets without overriding auto or sse fallback semantics", () => {
 	assert.equal(getEffectiveCodexTransport(undefined, undefined), "auto");
 	assert.equal(getEffectiveCodexTransport(undefined, { forceCachedWebSockets: true }), "auto");
@@ -196,6 +447,19 @@ test("parseSSE aborts response body reads when the caller aborts", async () => {
 	assert.equal(canceled, true);
 });
 
+test("parseSSE accepts CRLF chunks, joined data lines, and ignores done sentinel", async () => {
+	const response = new Response([
+		'data: {"type":"response.created",\r\n',
+		'data: "response":{"id":"resp_1"}}\r\n\r\n',
+		"data: [DONE]\r\n\r\n",
+	].join(""));
+
+	const events = [];
+	for await (const event of parseSSE(response)) events.push(event);
+
+	assert.deepEqual(events, [{ type: "response.created", response: { id: "resp_1" } }]);
+});
+
 test("getOpenAICodexImagePath saves images under the repo-local .pi/openai-codex-images directory", () => {
 	const filePath = getOpenAICodexImagePath("/repo", "resp_123", "ig_456", "png");
 	assert.equal(filePath, path.join("/repo", ".pi", "openai-codex-images", "ig_456-resp_123.png"));
@@ -215,6 +479,12 @@ test("getOpenAICodexImagePath falls back to png for unsafe image output formats"
 	const filePath = getOpenAICodexImagePath("/repo", "resp_123", "ig_456", "../../evil");
 	assert.equal(filePath, path.join("/repo", ".pi", "openai-codex-images", "ig_456-resp_123.png"));
 });
+
+test("getOpenAICodexImagePath sanitizes missing and unsafe identifiers", () => {
+	const filePath = getOpenAICodexImagePath("/repo", undefined, "ig weird/id", "webp");
+	assert.equal(filePath, path.join("/repo", ".pi", "openai-codex-images", "ig-weird-id-response.webp"));
+});
+
 test("buildGeneratedImageDisplayText surfaces the prompt and saved filename to the user", () => {
 	assert.equal(
 		buildGeneratedImageDisplayText({
@@ -349,6 +619,40 @@ test("activity dispatcher flushes queued display messages before shutdown clear"
 	assert.equal(sentMessages.length, 1);
 	assert.deepEqual(sentMessages[0]!?.options, { triggerTurn: false });
 	assert.equal((sentMessages[0]!?.message as { customType?: string }).customType, "codex-web-search-activity");
+});
+
+test("activity dispatcher preserves activity order and batches only adjacent web searches", async () => {
+	const sentMessages: Array<{ message: unknown; options: unknown }> = [];
+	const dispatcher = createActivityMessageDispatcher((message, options) => {
+		sentMessages.push({ message, options });
+	});
+	type SettledActivities = Parameters<typeof dispatcher.enqueueSettledActivities>[0];
+	const savedImage = {
+		absolutePath: "/repo/.pi/openai-codex-images/ig_1-resp_1.png",
+		relativePath: ".pi/openai-codex-images/ig_1-resp_1.png",
+		latestAbsolutePath: "/repo/.pi/openai-codex-images/latest.png",
+		latestRelativePath: ".pi/openai-codex-images/latest.png",
+		responseId: "resp_1",
+		callId: "ig_1",
+		outputFormat: "png",
+	};
+
+	dispatcher.enqueueSettledActivities([
+		{ kind: "web-search", search: { callId: "ws_1", queries: ["one"], sources: [] } },
+		{ kind: "web-search", search: { callId: "ws_2", queries: ["two"], sources: [] } },
+		{ kind: "image", savedImage, imageData: { data: "aW1hZ2U=", mimeType: "image/png" } },
+		{ kind: "web-search", search: { callId: "ws_3", queries: ["three"], sources: [] } },
+	] satisfies SettledActivities);
+	dispatcher.flushNow();
+
+	assert.equal(sentMessages.length, 3);
+	assert.equal((sentMessages[0]!.message as { customType?: string }).customType, WEB_SEARCH_ACTIVITY_MESSAGE_TYPE);
+	assert.deepEqual((sentMessages[0]!.message as { details?: { searches?: unknown[] } }).details?.searches?.map((search) => (search as { callId: string }).callId), ["ws_1", "ws_2"]);
+	assert.equal((sentMessages[1]!.message as { customType?: string }).customType, IMAGE_SAVE_DISPLAY_MESSAGE_TYPE);
+	assert.equal((sentMessages[2]!.message as { customType?: string }).customType, WEB_SEARCH_ACTIVITY_MESSAGE_TYPE);
+	assert.deepEqual((sentMessages[2]!.message as { details?: { searches?: unknown[] } }).details?.searches?.map((search) => (search as { callId: string }).callId), ["ws_3"]);
+	assert.deepEqual(sentMessages.map((entry) => entry.options), [{ triggerTurn: false }, { triggerTurn: false }, { triggerTurn: false }]);
+	assert.deepEqual(dispatcher.imagePreviewCache.get(savedImage.absolutePath), { data: "aW1hZ2U=", mimeType: "image/png" });
 });
 
 test("saveOpenAICodexGeneratedImage writes the decoded image bytes into the workspace-local cache", async () => {
